Massive amount of prototyping, Docker configuration, and refactoring.

+++ db/migrate/20160616191450_create_snomedct_descriptions.rb
@@ -2,11 +2,11 @@ class CreateSnomedctDescriptions < ActiveRecord::Migration
   def change
     create_table :snomedct_descriptions, id: :uuid do |t|
       t.uuid :snomedct_concept_id,	null: false
-      t.integer :snomedct_id,	null: false
+      t.bigint :snomedct_id,	null: false
       t.date :effective_time,	null: false
       t.boolean :active,	null: false
       t.bigint :module_id,	null: false
-      t.integer :concept_id,	null: false
+      t.bigint :concept_id,	null: false
       t.string :language_code,	null: false
       t.bigint :type_id,	null: false
       t.string :term,	null: false
